* Fix a couple of crashes in role path verbalization introduced in [1426] refs #395  * Display role numbers with duplicate replacement fields in reading shapes.  * Fixed issue with disabled context menu items for diagram tabs on secondary monitors to the left of or above the primary monitor.
SVN Revision: 1427

+++ ORMModel/Framework/Shell/MultiDiagramDocViewControl.cs
@@ -333,8 +333,10 @@ namespace ORMSolutions.ORMArchitect.Framework.Shell
 				if (contextMenuMessage = (m.Msg == WM_CONTEXTMENU))
 				{
 					uint lParam = (uint)m.LParam;
-					int x = (int)(lParam & 0xFFFF);
-					int y = (int)((lParam >> 16) & 0xFFFF);
+					// Convert to short instead of int to get the sign correct
+					// for multi-monitor systems with negative mouse positions.
+					int x = (short)(lParam & 0xFFFF);
+					int y = (short)((lParam >> 16) & 0xFFFF);
 					DiagramTabPage diagramTabPage = (x == -1 && y == -1) ? SelectedDiagramTab : GetTabAtPoint(base.PointToClient(new Point(x, y)));
 					myTrustDesignerForContextMenu = true;
 					myDesignerForContextMenu = (diagramTabPage != null) ? diagramTabPage.Designer : null;
